requirements.txt
# Core Framework
streamlit==1.31.0
streamlit-autorefresh

# Data Processing
pandas==2.1.4
numpy==1.26.3

# Firebase Integration
firebase-admin==6.4.0

# Visualization
plotly==5.18.0
matplotlib==3.8.2

# PDF Report Generation
reportlab==4.0.9

# API Requests (Groq)
requests==2.31.0

# Additional Utilities
python-dotenv==1.0.0
Pillow==10.2.0